+++ src/main/java/pisibg/model/dto/orderDTO/OrderEditResponseDTO.java
@@ -9,6 +9,7 @@ import pisibg.model.dto.productDTO.ProductOrderResponseDTO;
 import pisibg.model.pojo.Order;
 import pisibg.model.pojo.Product;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Map;
 import java.util.Queue;
@@ -25,9 +26,9 @@ public class OrderEditResponseDTO {
     private String address;
     private String paymentMethodType;
     private LocalDateTime createdAt;
-    private double grossValue;
-    private double discount;
-    private double netValue;
+    private BigDecimal grossValue;
+    private BigDecimal discount;
+    private BigDecimal netValue;
     @JsonProperty(value = "isPaid")
     private boolean isPaid;
