+++ src/main/java/pisibg/model/dto/orderDTO/OrderResponseDTO.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Component;
 import pisibg.model.dto.productDTO.ProductOrderResponseDTO;
 import pisibg.model.pojo.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Map;
 import java.util.Queue;
@@ -24,9 +25,9 @@ public class OrderResponseDTO {
     private String address;
     private String paymentMethodType;
     private LocalDateTime createdAt;
-    private double grossValue;
-    private double discount;
-    private double netValue;
+    private BigDecimal grossValue;
+    private BigDecimal discount;
+    private BigDecimal netValue;
     @JsonProperty(value = "isPaid")
     private boolean isPaid;
     private int totalProducts;
